backend/src/scripts/ml/requirements.txt
pandas>=2.2.0
numpy>=1.26.0
xgboost>=2.1.0
scikit-learn==1.6.1
tensorflow-cpu>=2.16.0
joblib>=1.3.0
prophet
